// oneHub/datart-versions/datart-1.0.0-rc.1/frontend/src/app/pages/MainPage/pages/ViewPage/slice/types.ts
export interface Schema {
  name: string;
  primaryKey?: boolean;
  type: DataViewFieldType;
}

export enum ColumnRole {
  Role = 'role',
  Hierarchy = 'hierachy',
  Table = 'table',
}

export interface Column extends Schema {
  category?: ColumnCategories;
  index?: number;

  role?: ColumnRole;
  children?: Column[];
  path?: string[];
  displayName?: string;
}